+++ src/main/java/com/hubspot/jinjava/el/ext/NamedParameterOperator.java
@@ -1,5 +1,7 @@
 package com.hubspot.jinjava.el.ext;
 
+import com.hubspot.jinjava.interpret.InterpretException;
+
 import de.odysseus.el.tree.impl.Parser.ExtensionHandler;
 import de.odysseus.el.tree.impl.Parser.ExtensionPoint;
 import de.odysseus.el.tree.impl.Scanner;
@@ -13,6 +15,9 @@ public class NamedParameterOperator {
   public static final ExtensionHandler HANDLER = new ExtensionHandler(ExtensionPoint.ADD) {
     @Override
     public AstNode createAstNode(AstNode... children) {
+      if (!(children[0] instanceof AstIdentifier)) {
+        throw new InterpretException("Expected IDENTIFIER, found " + children[0].toString());
+      }
       AstIdentifier name = (AstIdentifier) children[0];
       return new AstNamedParameter(name, children[1]);
     }